# doc/technical/openapi.yaml
openapi: 3.1.0
info:
  title: 'SoManAgent API'
  version: 1.0.0
  description: 'Versioned OpenAPI contract for the SoManAgent backend. This specification is spec-first: operations may be documented ahead of implementation when they explicitly set `x-somanagent-implemented` to `false`.'
servers:
  - url: 'http://localhost:8080'
...
paths:
  /api/agents:
    get:
      tags:
        - Agents
      summary: 'List agents'
      operationId: listAgents
      responses:
        200:
          $ref: '#/components/responses/Ok'
    post:
      tags:
        - Agents
      summary: 'Create an agent'
      operationId: createAgent
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        201:
          $ref: '#/components/responses/Created'
  '/api/agents/{id}':
    get:
      tags:
        - Agents
      summary: 'Get an agent'
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    patch:
      tags:
        - Agents
      summary: 'Update an agent'
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    delete:
      tags:
        - Agents
      summary: 'Delete an agent'
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
  '/api/agents/{id}/status':
    get:
      tags:
        - Agents
      summary: 'Get agent status'
      operationId: getAgentStatus
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
  '/api/agents/connectors/{connector}/models':
    get:
      tags:
        - Agents
      summary: 'List normalized models for an agent connector'
      operationId: listAgentConnectorModels
      parameters:
        - $ref: '#/components/parameters/ConnectorPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  '/api/agents/{id}/executions':
    get:
      tags:
        - Agents
      summary: 'List executions for an agent'
      operationId: listAgentExecutions
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
  '/api/roles/{id}':
    get:
      tags:
        - Roles
      summary: 'Get a role'
      operationId: getRole
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    patch:
      tags:
        - Roles
      summary: 'Update a role'
      operationId: updateRole
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    delete:
      tags:
        - Roles
      summary: 'Delete a role'
      operationId: deleteRole
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
...
  '/api/teams/{id}/agents':
    post:
      tags:
        - Teams
      summary: 'Attach an agent to a team'
      operationId: addTeamAgent
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  '/api/teams/{id}/agents/{agentId}':
    delete:
      tags:
        - Teams
      summary: 'Detach an agent from a team'
      operationId: removeTeamAgent
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/AgentIdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
...
  '/api/tickets/{id}':
    get:
      tags:
        - Tickets
      summary: 'Get a ticket'
      operationId: getTicket
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        200:
          description: 'Successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
    patch:
      tags:
        - Tickets
      summary: 'Update a ticket'
      operationId: updateTicket
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          description: 'Successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
    delete:
      tags:
        - Tickets
      summary: 'Delete a ticket'
      operationId: deleteTicket
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
...
  '/api/tokens/agents/{agentId}':
    get:
      tags:
        - Tokens
      summary: 'Get token usage for one agent'
      operationId: getAgentTokenUsage
      parameters:
        - $ref: '#/components/parameters/AgentIdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AgentIdPath:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ConnectorPath:
      name: connector
      in: path
      required: true
      schema:
        type: string
...
  requestBodies:
    ArbitraryJsonBody:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArbitraryJsonObject'
...
  responses:
    Ok:
      description: 'Successful response'
    Created:
      description: 'Resource created'
...
    NoContent:
      description: 'Resource deleted successfully'
  schemas:
...
    ArbitraryJsonObject:
      type: object
      additionalProperties: true
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        type:
          type: string
          enum: [user_story, bug]
        title:
          type: string
        description:
          type: string
          nullable: true
        initialRequest:
          type: string
          nullable: true
          description: 'Original description entered by the user at ticket creation time, before any agent rewrites it.'
        initialTitle:
          type: string
          nullable: true
          description: 'Original title entered by the user at ticket creation time, before any agent rewrites it.'
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - title
        - status
        - priority
        - progress
        - createdAt
        - updatedAt